SmartTicket/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.5,<0.0.6
python-dotenv>=0.19.0,<0.20.0
email-validator>=1.1.3,<2.0.0

# OpenAI and LangChain
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Vector Store and Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Testing
pytest>=6.2.4,<7.0.0
requests>=2.26.0,<3.0.0
httpx>=0.18.2,<0.19.0 